api/requirements.txt
azure-functions
azure-data-tables==12.5.0
requests
